+++ gamification-specs/src/test/java/ch/heigvd/gamification/api/spec/steps/UserSteps.java
@@ -1,25 +1,11 @@
 package ch.heigvd.gamification.api.spec.steps;
 
 import ch.heigvd.gamification.ApiException;
-import ch.heigvd.gamification.ApiResponse;
 import ch.heigvd.gamification.api.DefaultApi;
-import ch.heigvd.gamification.api.dto.ApiKey;
-import ch.heigvd.gamification.api.dto.Application;
-import ch.heigvd.gamification.api.dto.Badge;
-import ch.heigvd.gamification.api.dto.Event;
-import ch.heigvd.gamification.api.dto.EventEventparams;
 import ch.heigvd.gamification.api.dto.User;
 import ch.heigvd.gamification.api.spec.helpers.Environment;
-import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-
-import java.util.List;
-import java.util.UUID;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 
 public class UserSteps {
 
@@ -29,16 +15,7 @@ public class UserSteps {
     private BasicSteps basicSteps;
     private ApplicationSteps applicationSteps;
 
-    User user;
-
-
-    private ApiResponse lastApiResponse;
-    private ApiException lastApiException;
-    private boolean lastApiCallThrewException;
-    private int lastStatusCode;
-
-    private String lastReceivedLocationHeader;
-    private User lastReceivedUser;
+    private User user;
 
     public  UserSteps(Environment environment, BasicSteps basicSteps, ApplicationSteps applicationSteps) {
         this.environment = environment;
@@ -48,7 +25,7 @@ public class UserSteps {
     }
 
     @Given("I have a user payload")
-    public void i_have_a_user_payload() throws Throwable {
+    public void i_have_a_user_payload() {
         user = new ch.heigvd.gamification.api.dto.User()
                 .id(37)
                 .username("Jean");
@@ -56,7 +33,7 @@ public class UserSteps {
     }
 
     @When("^I POST the user payload to the /users endpoint$")
-    public void i_POST_the_user_payload_to_the_users_endpoint() throws Throwable {
+    public void i_POST_the_user_payload_to_the_users_endpoint() {
         try {
             basicSteps.processApiResponse(api.createUserWithHttpInfo(applicationSteps.getApiKey(), user));
         } catch (ApiException e) {
